[PATCH] Fix d_u32align probe on ARM
From: Marc Pignat <marc@pignat.org>
Date: Wed, 23 Jun 2010 15:38:37 +0200

My proposed solution is to use the 'volatile' keyword to make sure the
compiler will really do the job.

Please note that it does not work with -O3, but the configure script
compile with -O2, so this should be fine.

At this time I've only tested the Configure scripts, which now detects
the align problem on ARM. I'm compiling the whole thing to check if the
md5 bug goes away!

I hope this will fix this *5 years* old bug, and the upstream bug as
well!

Signed-off-by: H.Merijn Brand <merijn@nb09.procura.nl>