Fix [perl #116523] Configure is confused by '=' in output from uname -a
Previously, Configure used to try to post-process the output of uname -a.
This was to support Xenix, which gave output like this:

     sysname=XENIX
     nodename=whatever
     release=2.3.2 .. etc...

Configure attempted to parse that output and build up a $myuname variable
that resembled those generated on traditional Unix systems.

More recently, a user reported uname -a output on a CVS checkout of
NetBSD that also included information such as

    root=username@hostname:/dirname module=src branch=HEAD

The Configure code misparsed that and ended up throwing most of it away.

This patch takes the simplest approach and just skips the attempt to
process the Xenix output.