Move the thrice-repeated "64bitint is buggy?" test from hints to Configure.
Backport 16e6698bf53cbe5099ca8f3

OpenBSD and its two* forks each have the same "is 64 bit int support buggy?"
test code as a callback unit duplicated in their hints files. As the code is
portable C, there seems to be no harm in moving it to Configure, and running
the test on all platforms. This reduces code duplication, and will reduce it
further if another OpenBSD fork appears.

* This week.