More specific comment based on findings in RT#7586
Suggested by Nicholas Clark, somewhat tidied to Glossary standards

[perl #7586] STDOUT is not flushed before exec().

Nicholas comment in the original fix:

It turns out* that $Config{fflushNULL} on HP-UX is *STILL* undef, because
fflush(NULL) wrongly closes STDIN if STDIN is a pipe (which is what the
Configure test probes for).

I wonder if HP are even aware that their platform is still not quite 100%
C89 conformant? Of if they'll fix it in time for the 25th anniversary of
the first ANSI C standard?

(Also, given that that this bug had been present on Solaris, and on early
FreeBSDs, I wonder if it's a bug in common code they inherit from, or just
convergent mistake making)

Nicholas Clark

* At least on the system Merijn provides for us.