Handle hyphens when searching cpp symbols
Simplified solution to a problem signaled and patched by Brian

This is really just for Android, which in Configure is
known as linux-android; the hyphen is passes as-is to
the preprocessor, which handles it as a minus.

So don't do that.

Original patch was ecd78e215c9b4ab3f3ab71b30e76bbead372a84e
Author: Brian Fraser <fraserbn@gmail.com>  2014-07-25 19:46:41