Use size_t and ssize_t directly
Backport of this commit, which is currently on the smoke-me/ilmari/size_t
branch in perl5.git:

commit 33bf7414d386dc2895cc0ef78e40fc0667b52262
Author: Dagfinn Ilmari Mannsåker <ilmari@ilmari.org>
Date:   Fri Feb 26 17:42:48 2016 +0000

    Use (s)size_t constitently

    size_t exists everywhere, but not ssize_t, so keep the probing for
    that (but call the type ssize_t, not SSize_t).

    Keep the old typenames for non-PERL_CORE.

In fact, this goes further: the evidence I can find is that all supported
platforms (including both HP/UX 10 and AIX) provide ssize_t, so that probe
is also omitted.

Ilmari's change retains unconditional definitions of the capitalised names
in perl.h, in a block that's compiled only outside PERL_CORE, for the
benefit of XS modules that expect to be able to use them. The consequence of
this is that metaconfig's scan treats those symbols as in use, and so pulls
in the relevant units. We handle this by deleting most of the contents of
the sizetype.U and ssizetype.U units; in particular, they depend on nothing,
and have no ?MAKE commands to actually trigger the insertion of the probes
into the generated files.