For Perl, assume C89 prototypes exist
Perl now relies on C89, including prototypes. However, it must continue to
define CAN_PROTOTYPE for the benefit of XS modules. So the Perl-specific
modified version of U/modified/prototype.U is now a stub that probes for
nothing, but still overrides the dist/ version of prototype.U (which would
otherwise find the remaining CAN_PROTOTYPE occurrences in the core and
include the probe anyway).

In addition, the Protochk unit itself looks at CAN_PROTOTYPE, and is used by
many other units. So create a modified version which doesn't do that.