contains.U: Need a path to the generated script
I tested what would happen with 'contains' if grep is actually broken on
the box, and it failed because it didn't know where the generated
work-around script is.

This commit fixes this, but the fact that no failures have been showing
up indicates that modern greps just aren't broken this way.