d_setlocale.U: Pay attention to hints file
This was ignoring any hints file setting of d_setlocale, and checking if
the system appears to have a sane one.  But there are cases where the
setlocale isn't really sane, even though it appears to be, and the hints
file is used to override this setting.  This commit now pays attention
to that override.