Trap SIGBUS when Configure checks for va_copy (#18148)
On several systems the attempt to determine if we &quot;need va_copy() or
similar&quot; results in a SIGBUS instead of a SIGSEGV because the fault
happens pointing at an io device address.

While reviewing accounting logs during the OpenBSD system build, bluhm@
noticed a core dump and deraadt@ tracked it down.

While researching it was discovered that this is a fairly common case in
other systems as well.

Original diff by deraadt@