only #include <xlocale.h> when it is actually needed
Backport of 68f1b6d700b9662cfdec43ceddbd086d4b59c87e

This header was originally only needed for builds on darwin and
FreeBSD, but was being included whenever it was detected.

This has caused problems when what was an internal header was
removed (from glibc) and in general wasn't needed anyway.

On FreeBSD only localeconv_l() requires xlocale.h, so we test
specifically for that.