properly populate osvers on Dragonfly BSD when the hostname isn't set
Backport of be2d063f33e907f7892a8ff7e6f73cf7aa86397a
With additional changes to leading whitespace to make the file consistent again

A default installation of DragonflyBSD doesn't necessarily set the
hostname, resulting in a `uname -a` like:

  DragonFly  6.2-RELEASE DragonFly v6.2.2-RELEASE ...

which resulted in osvers being set to "dragonfly", which isn't
especially useful.

So check that $3 is numeric in some sense before using it, falling
back to $2 if it isn't.  This should only happen when the hostname
isn't set.