Remove test for thread-safe nl_langinfo_l()
Testing has shown that it was naive of me to think that this function
could be easily tested for thread-safety.  What actually was instead
tested for was if the returned buffer is per-thread or global.  There
may very well be not-as-obvious interactions with other threads.  And on
some platforms that turned out to be the case.  Better to use plain
nl_langinfo, which has the mutexes to prevent collisions.

This commit combines the tests for these into the Unit for plain
nl_langinfo.