Add probe to find syntax of disparate LC_ALL
Not all locale categories need be set to the same locale.  The syntax of
LC_ALL under this condition varies by platform.  This commit adds a
probe to determine what this platform uses, and removes unused code from
another unit that was to be the beginning of probing for the same
information.  I realized that a separate probe was needed.